feat(core): export buildResolvedEntries for entry-list loaders
extract the batched entry resolver (raw entry rows -> ResolvedEntry with
author, terms, and permalink) out of resolve.ts into its own module and
export it from the core barrel. plugins building custom entry lists — a
blog 'related posts' dep, popular/curated strips — need the same render-ready
hydration the front-page and archive paths use, without duplicating the
two-query batch. pure move: the resolver body is unchanged.
